pinned file dropdown in separate component and field css done
