+++ client/src/containers/file-source-radio/index.tsx
@@ -1,35 +1,21 @@
 import React from 'react';
 import FileSourceRadio from '../../components/file-source-radio';
 import { connect } from 'react-redux';
-import { selectDocument, selectPinnedFile, setFileSource } from '../../state/actions/document';
+import { selectDocument, setFileSource } from '../../state/actions/document';
 import { showError, hideError } from '../../state/actions/alert';
 import { setIsLoading } from '../../state/actions/loading';
-import { createJSONFromSpreadsheet, getPinnedFile } from '../../utils/spreadsheet.utils';
-import { UserState } from '../../state/reducers/user';
 
 interface Props {
-  showError: any;
-  hideError: any;
-  setIsLoading: any;
   selectDocument: any;
   setFileSource: any;
-  selectedPinnedFileId: any;
   fileSource: string;
-  selectPinnedFile: any;
-  pinnedFiles: Array<any>;
   index: number;
 }
 
 const FileSourceRadioContainer = ({
-  showError,
-  hideError,
-  setIsLoading,
   selectDocument,
-  selectedPinnedFileId,
   setFileSource,
-  selectPinnedFile,
   fileSource,
-  pinnedFiles,
   index
 }: Props) => {
 
@@ -40,34 +26,10 @@ const FileSourceRadioContainer = ({
     }
   }
 
-  const handlePinnedFileSelection = async (event: any) => {
-    console.log('event', event)
-    const fileId: string = event.target.value;
-    selectPinnedFile(fileId);
-    setIsLoading(true);
-    try {
-      const pinnedFileBlob = await getPinnedFile(fileId);
-      const wsDataObj: Array<any> = await createJSONFromSpreadsheet(pinnedFileBlob.data);
-      selectDocument(wsDataObj, index, fileId, pinnedFileBlob.data);
-      hideError();
-      setIsLoading(false);
-    } catch (err: any) {
-      console.log('err', err);
-      showError('Failed to get pinned file for user.');
-      setIsLoading(false);
-      setTimeout(() => {
-        hideError();
-      }, 5000);
-    }
-  }
-
   return (
     <FileSourceRadio
       handleFileTypeChange={handleFileTypeChange}
-      handlePinnedFileSelection={handlePinnedFileSelection}
       fileSource={fileSource}
-      selectedPinnedFileId={selectedPinnedFileId}
-      pinnedFiles={pinnedFiles}
       index={index}
     />
   );
@@ -84,7 +46,6 @@ const mapDispatchToProps = (dispatch: any) => ({
   setIsLoading: (isLoading: boolean) => dispatch(setIsLoading(isLoading)),
   selectDocument: (wsDataObj: any, index: number, filename: string, fileBlob: any) => dispatch(selectDocument(wsDataObj, index, filename, fileBlob)),
   setFileSource: (index: number, event: any) => dispatch(setFileSource(index, event)),
-  selectPinnedFile: (pinnedFileId: string) => dispatch(selectPinnedFile(pinnedFileId))
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(FileSourceRadioContainer);
